Get configuration object

`getConfig` returns a config object required to use ShinyItemBuilder.

Usage

getConfig(
  WindowTitle = "MyAssessment",
  Verbose = FALSE,
  WWWfolder = "_mywww",
  Datafolder = "_mydata",
  posH = "center",
  posV = "center",
  scaling = "scale-up-down",
  sessiontype = "sessionstorage",
  maintenancePassword = "",
  maintenanceQuery = "maintenance",
  maintenanceKey = list(key = "x", ctrl = T, shift = F, alt = F),
  clearPreviousTaskState = TRUE
)

Arguments

WindowTitle

Title of the html browser window.

Verbose

Should the package provide log information to the console?

WWWfolder

Folder to store html/javascript files required to run the assessment

Datafolder

Folder to store data files for started/completed assessments

posH

Horizontal orientation (should be one of 'left', 'right' or 'center')

posV

Vertical orientation (should be one of 'top', 'bottom' or 'center')

scaling

Scaling of content (should be one of 'scale-up', 'scale-down', 'scale-up-down' or 'no-scaling')

sessiontype

Session storage (should be one of 'sessionstorage', 'cookie', 'localstorage' or 'provided' )

maintenancePassword

Password to access data online (no access possible if not defined).

maintenanceQuery

Query string parameter name to access data online.

maintenanceKey

Keyboard shortcut to access the maintenance dialog (provide a list in the following form: `list(key="X", ctrl=T, shift=F, alt=F)`)

clearPreviousTaskState

If your test doesn't allow backward navigation, let Itembuilder's runtime flush previous task states when navigating forward

Value

config object (list)

Details

The returned object is a list that contains the different configurations for an assessment using ShinyItemBuilder.

Examples

if (FALSE) { # \dontrun{
 conf <- getConfig()
} # }